fix: preserve architecture ratchets for style sets
