+++ apps/api/src/handlers/style-sets/update.ts
@@ -3,15 +3,12 @@ import { and, eq } from "drizzle-orm";
 import { t } from "elysia";
 
 import { styleSets } from "@/api/db/schema";
-import {
-  styleSetColumns,
-  styleSetExportFileName,
-} from "@/api/handlers/style-sets/shared";
 import { createSafeRootHandler } from "@/api/lib/api-handlers";
 import type { HandlerConfig } from "@/api/lib/api-handlers";
 import { AUDIT_ACTION, AUDIT_RESOURCE_TYPE } from "@/api/lib/audit-log";
 import { tDefaultVarchar, tSafeId } from "@/api/lib/custom-schema";
 import { HandlerError } from "@/api/lib/errors/tagged-errors";
+import { styleSetColumns, styleSetExportFileName } from "@/api/lib/style-sets";
 
 const paramsSchema = t.Object({ styleSetId: tSafeId("styleSet") });
 const bodySchema = t.Object({ name: tDefaultVarchar });
