+++ apps/api/src/handlers/templates/create-blank.test.ts
@@ -9,7 +9,7 @@ import {
 
 import { discoverTemplate } from "@/api/handlers/docx/discover-template";
 import { readManifest } from "@/api/handlers/docx/template-manifest";
-import { createTemplateBuffer } from "@/api/handlers/templates/create-template-buffer";
+import { createTemplateBuffer } from "@/api/lib/create-template-buffer";
 
 // The blank-create handler's only logic on top of the shared
 // `createStoredTemplate` recipe is generating the source buffer from a
