+++ apps/api/src/handlers/templates/create-from-styles.ts
@@ -2,7 +2,6 @@ import { Result } from "better-result";
 import { t } from "elysia";
 
 import type { SafeDb } from "@/api/db/safe-db";
-import { createTemplateBuffer } from "@/api/handlers/templates/create-template-buffer";
 import {
   type CreatedTemplate,
   createStoredTemplate,
@@ -14,6 +13,7 @@ import type {
 } from "@/api/lib/api-handlers";
 import type { AuditRecorder } from "@/api/lib/audit-log";
 import type { SafeId } from "@/api/lib/branded-types";
+import { createTemplateBuffer } from "@/api/lib/create-template-buffer";
 import { tDefaultVarchar } from "@/api/lib/custom-schema";
 import { HandlerError } from "@/api/lib/errors/tagged-errors";
 import { FILE_SIZE_LIMITS } from "@/api/lib/limits";
