+++ apps/web/src/routes/_protected.knowledge/-queries.ts
@@ -11,7 +11,6 @@ import { toSafeId } from "@/lib/safe-id";
 
 const SKILLS_PAGE_SIZE = 100;
 const PLAYBOOKS_PAGE_SIZE = 50;
-const STYLE_SETS_PAGE_SIZE = 100;
 
 type SkillsPageKey = {
   limit: number;
@@ -93,13 +92,6 @@ export const knowledgeKeys = {
   templateRecipes: {
     all: (organizationId: string) => ["template-recipes", organizationId],
   },
-  styleSets: {
-    all: (organizationId: string) => ["style-sets", organizationId],
-    list: (organizationId: string) => [
-      ...knowledgeKeys.styleSets.all(organizationId),
-      "list",
-    ],
-  },
   clauses: {
     all: (organizationId: string) => ["clauses", organizationId],
     list: (
@@ -152,22 +144,6 @@ export const knowledgeKeys = {
     ],
   },
 };
-
-export const styleSetsOptions = (organizationId: string) =>
-  queryOptions({
-    queryKey: knowledgeKeys.styleSets.list(organizationId),
-    queryFn: async ({ signal }) => {
-      const response = await api["style-sets"].get({
-        query: { limit: STYLE_SETS_PAGE_SIZE },
-        fetch: { signal },
-      });
-      if (response.error) {
-        throw toAPIError(response.error);
-      }
-      return response.data;
-    },
-    staleTime: STALE_TIME.FIVE.MINUTES,
-  });
 
 // ── Template queries ────────────────────────────────
